complete pipeline and models

+++ corpora/google.py
@@ -31,20 +31,24 @@ def arrange():
             continue
         os.makedirs(os.path.join(f, 'wavs'), exist_ok=True)
         for wav in glob.iglob(os.path.join(f, '*.wav')):
-            shutil.move(wav, os.path.join(f, 'wavs'))
+            if not os.path.exists(os.path.join(f, 'wavs', os.path.split(wav)[-1])):
+                shutil.move(wav, os.path.join(f, 'wavs'))
 
 def merge():
     base_path = os.path.join(dataset_path, 'google')
     for f in glob.iglob(os.path.join(base_path, '*')):
-        if not f.endswith('male'):
+        if not f.endswith('male') and not f.endswith('female'):
             continue
         lang_name = os.path.split(f)[-1][:5]
         out_path = os.path.join(base_path, lang_name)
-        shutil.move(os.path.join(f, 'wavs'), out_path)
+        os.makedirs(os.path.join(out_path, 'wavs'), exist_ok=True)
+        for wav in glob.iglob(os.path.join(f, 'wavs', '*.wav')):
+            if not os.path.exists(os.path.join(out_path, 'wavs', os.path.split(wav)[-1])):
+                shutil.move(wav, os.path.join(out_path, 'wavs'))
         lines = open(os.path.join(f, "line_index.tsv"), "r", encoding='utf-8').read().splitlines()
         open(os.path.join(out_path, "line_index.tsv"), 'a', encoding="utf-8").writelines(
             [l + '\n' for l in lines]
-        )
+        ) # Warn: must only run once, or there will be duplicates
 
 def process(base_path):
     lang = os.path.split(base_path)[-1]
@@ -53,9 +57,10 @@ def process(base_path):
     wav_output_path = os.path.join(output_path, 'wavs')
     os.makedirs(wav_output_path, exist_ok=True)
 
-    samples = open(os.path.join(base_path, "line_index.tsv"), encoding='utf-8').read().splitlines()
+    samples = open(os.path.join(base_path, "si_lk.lines.txt" if lang == "si_lk" else "line_index.tsv"),
+                   encoding='utf-8').read().splitlines()
     n_skip = 0
-    print(lang, '\n', samples[0], '\n', samples[1], '\n', samples[2])
+    print(lang, '\n', samples[0], '\n', samples[1], '\n', samples[2], '\n')
 
     for sample in tqdm.tqdm(samples):
         if lang == "si_lk":
@@ -68,10 +73,9 @@ def process(base_path):
             continue
         if name.endswith('.wav'):
             name = name[:-4]
+        if script[-2:] == "\\n":
+            script = script[:-2]
         words = script.split(' ')
-        if len(words) <= 2:
-            n_skip += 1
-            continue
         words = [w for w in words if w]
         for k, w in enumerate(words):
             if w[0] == "[" and w[-1] == "]":
@@ -88,9 +92,8 @@ def process(base_path):
         spk_name = name.split('_')[0] + name.split('_')[1]
         wav_file = os.path.join(base_path, 'wavs', name + '.wav')
         dur = librosa.get_duration(filename=wav_file)
-        if not dur <= 20 or any([c in "1234567890" for c in script]):
-            if n_skip < 10:
-                print(dur, script)
+        if any([c in "1234567890" for c in script]):
+            print(name, dur, script)
             n_skip += 1
             continue
         spk_samples[spk_name].append((wav_file, script, dur))
@@ -98,6 +101,7 @@ def process(base_path):
     n_spk_skip = 0
     for spk in list(spk_samples.keys()):
         if len(spk_samples[spk]) < 50:
+            print(spk, len(spk_samples[spk]))
             n_skip += len(spk_samples[spk])
             del spk_samples[spk]
             n_spk_skip += 1
@@ -123,9 +127,12 @@ def process(base_path):
     )
 
 if __name__ == '__main__':
-    # merge()
-    # arrange()
+    extract()
+    arrange()
+    merge()
     for f in glob.iglob(os.path.join(dataset_path, 'google', '*')):
         if not os.path.isdir(f) or len(os.path.split(f)[-1]) != 5:
+            continue
+        if os.path.exists(os.path.join(transformed_path, 'google_' + os.path.split(f)[-1])):
             continue
         process(f)
